+++ tests/testthat/test-compile_pseudo_pop.R
@@ -11,7 +11,7 @@ test_that("Compiling pseudo pop works as expected.", {
   # Wrong ci_appr
   expect_error(compile_pseudo_pop(data_obj = gps_obj,
                                   ci_appr = "grounding",
-                                  gps_model = "parametric",
+                                  gps_density = "normal",
                                   bin_seq = NULL,
                                   exposure_col_name = c("w"),
                                   nthread = 1,
@@ -26,7 +26,7 @@ test_that("Compiling pseudo pop works as expected.", {
   set.seed(509)
   pseudo_pop_1 <- compile_pseudo_pop(data_obj = gps_obj,
                                      ci_appr = "matching",
-                                     gps_model = "parametric",
+                                     gps_density = "normal",
                                      bin_seq = NULL,
                                      exposure_col_name = c("w"),
                                      nthread = 1,
@@ -52,7 +52,7 @@ test_that("Compiling pseudo pop works as expected.", {
   obj$dataset$id <- obj$dataset$row_index
   pseudo_pop_2 <- compile_pseudo_pop(data_obj = obj,
                                      ci_appr="weighting",
-                                     gps_model="parametric",
+                                     gps_density = "normal",
                                      bin_seq = NULL,
                                      exposure_col_name = c("w"),
                                      nthread = 1,